pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-applemusic-api"
version = "0.1.0"
description = "MCP server for Apple Music playlist management via REST API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Eric Pheterson" }
]
keywords = ["mcp", "apple-music", "musickit", "playlist", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "mcp>=1.0.0",
    "pyjwt[crypto]>=2.8.0",
    "requests>=2.31.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
applemusic-mcp = "applemusic_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/epheterson/mcp-applemusic-api"
Repository = "https://github.com/epheterson/mcp-applemusic-api"
Issues = "https://github.com/epheterson/mcp-applemusic-api/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/applemusic_mcp"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
